feat: Update ExerciseCompleteResponse and ExerciseQuestionResponse for improved data structure and accuracy calculation

+++ src/main/java/com/kiovant/englishme/dto/ExerciseCompleteResponse.java
@@ -9,7 +9,7 @@ public record ExerciseCompleteResponse(
         int totalQuestions,
         int correct,
         int incorrect,
-        int accuracyPercent,
+        double accuracyPercent,
         int xpEarned,
         long totalXp,
         int dailyEarnedXp,
